DB-3176: separating getting menu logic from pages
