+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules
 .pnpm-debug.log
 .pnpm-store
+.DS_Store
