+++ starters/next-wordpress-starter/pages/pages/index.js
@@ -1,8 +1,7 @@
 import { NextSeo } from "next-seo";
 import Layout from "../../components/layout";
 import PageHeader from "../../components/page-header.js";
-import { client } from "../../lib/WordpressClient";
-import { gql } from "@pantheon-systems/wordpress-kit";
+import { getFooterMenu } from "../../lib/Menus";
 
 export default function PageListTemplate({ menuItems }) {
   return (
@@ -24,38 +23,11 @@ export default function PageListTemplate({ menuItems }) {
 }
 
 export async function getStaticProps() {
-  const query = gql`
-    query FooterMenuQuery {
-      menus(where: { location: FOOTER }) {
-        edges {
-          node {
-            id
-            menuItems {
-              edges {
-                node {
-                  id
-                  uri
-                  label
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  `;
-
-  const data = await client.request(query);
-
-  const footerMenu = data.menus.edges.flatMap((node) => {
-    return node.node.menuItems.edges.map((menuItem) => {
-      return menuItem.node;
-    });
-  });
+  const menuItems = await getFooterMenu();
 
   return {
     props: {
-      menuItems: footerMenu,
+      menuItems,
     },
   };
 }
